spike: renaming sqliteutils to sqlitetypeutils
