+++ packages/engine-server/src/drivers/sqlite/index.ts
@@ -1,6 +1,6 @@
 export * from "./SqliteDbFactory";
 export * from "./SqliteMetadataStore";
 export * from "./tables";
-export * from "./SqliteUtils";
+export * from "./SqliteTypeUtils";
 export * from "./SqliteQueryUtils";
 export * from "./SqliteError";
