+++ packages/engine-server/src/drivers/sqlite/tables/LinksTableUtils.ts
@@ -4,7 +4,7 @@ import { ResultAsync } from "neverthrow";
 import { Database } from "sqlite3";
 import { SqliteError } from "../SqliteError";
 import { SqliteQueryUtils } from "../SqliteQueryUtils";
-import { getIntegerString, getSQLValueString } from "../SqliteUtils";
+import { getIntegerString, getSQLValueString } from "../SqliteTypeUtils";
 
 // TODO Should be an enum...
 export type LinkType =
